requirements.txt
Django==1.10.4
django-filter==1.0.1
djangorestframework==3.5.3
python-dateutil==2.6.0
six==1.10.0
django-cors-headers==1.1.0